+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- Make `pickupSLAs` query return all pickup points close to location sent, even uunavailable pickups.
+- Make `pickupSLAs` query return all pickup points close to location sent, even unavailable ones.
 
 ## [2.113.6] - 2020-01-28
 ### Removed